+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.ali.fitnessfirst"
+        applicationId "com.example.fitnessfirst"
         minSdk 26
         targetSdk 32
         versionCode 1
