+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ali.fitnessfirst">
+    package="com.example.fitnessfirst">
 
     <application
         android:allowBackup="true"
@@ -10,40 +10,40 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.FitnessFirst">
         <activity
-            android:name=".home.ui.exercise.activities.ShouldersExerciseActivity"
+            android:name="com.example.fitnessfirst.home.ui.exercise.activities.ShouldersExerciseActivity"
             android:exported="false" />
         <activity
-            android:name=".home.ui.exercise.activities.LegsExerciseActivity"
+            android:name="com.example.fitnessfirst.home.ui.exercise.activities.LegsExerciseActivity"
             android:exported="false" />
         <activity
-            android:name=".home.ui.exercise.activities.CoreExerciseActivity"
+            android:name="com.example.fitnessfirst.home.ui.exercise.activities.CoreExerciseActivity"
             android:exported="false" />
         <activity
-            android:name=".home.ui.exercise.activities.ArmsExerciseActivity"
+            android:name="com.example.fitnessfirst.home.ui.exercise.activities.ArmsExerciseActivity"
             android:exported="false" />
         <activity
-            android:name=".home.ui.exercise.activities.BackExerciseActivity"
+            android:name="com.example.fitnessfirst.home.ui.exercise.activities.BackExerciseActivity"
             android:exported="false" />
         <activity
-            android:name=".home.ui.exercise.activities.ChestExerciseActivity"
+            android:name="com.example.fitnessfirst.home.ui.exercise.activities.ChestExerciseActivity"
             android:exported="false" />
         <activity
-            android:name=".login.ForgotPasswordActivity"
+            android:name="com.example.fitnessfirst.login.ForgotPasswordActivity"
             android:exported="false"
             android:label="@string/title_activity_forgot_password"
             android:theme="@style/Theme.FitnessFirst.NoActionBar" />
         <activity
-            android:name=".home.HomeActivity"
+            android:name="com.example.fitnessfirst.home.HomeActivity"
             android:exported="false"
             android:label="@string/title_activity_home" />
         <activity
-            android:name=".login.LoginActivity"
+            android:name="com.example.fitnessfirst.login.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".login.SignUpActivity"
+            android:name="com.example.fitnessfirst.login.SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".LoadingActivity"
+            android:name="com.example.fitnessfirst.LoadingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
