--- app/src/main/java/com/ali/fitnessfirst/LoadingActivity.java
@@ -1,25 +1,0 @@
-package com.ali.fitnessfirst;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-
-import com.ali.fitnessfirst.login.LoginActivity;
-
-public class LoadingActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_loading);
-        findViewById(R.id.btn_started).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent =  new Intent(LoadingActivity.this, LoginActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-}
