--- app/src/main/java/com/ali/fitnessfirst/home/ui/exercise/activities/LegsExerciseActivity.java
@@ -1,16 +1,0 @@
-package com.ali.fitnessfirst.home.ui.exercise.activities;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import com.ali.fitnessfirst.R;
-
-public class LegsExerciseActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_legs_exercise);
-    }
-}
