--- app/src/main/java/com/ali/fitnessfirst/home/ui/exercise/activities/ShouldersExerciseActivity.java
@@ -1,18 +1,0 @@
-package com.ali.fitnessfirst.home.ui.exercise.activities;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import com.ali.fitnessfirst.R;
-
-public class ShouldersExerciseActivity extends AppCompatActivity {
-
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_shoulders_exercise);
-    }
-}
