--- app/src/main/java/com/ali/fitnessfirst/model/Todo.java
@@ -1,10 +1,0 @@
-package com.ali.fitnessfirst.model;
-
-
-public class Todo {
-
-    private String name;
-
-
-
-}
