+++ app/src/main/java/com/example/fitnessfirst/LoadingActivity.java
@@ -1,0 +1,25 @@
+package com.example.fitnessfirst;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.fitnessfirst.login.LoginActivity;
+
+public class LoadingActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_loading);
+        findViewById(R.id.btn_started).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent =  new Intent(LoadingActivity.this, LoginActivity.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
