+++ app/src/main/java/com/example/fitnessfirst/home/ui/bmi/NotificationsViewModel.java
@@ -1,0 +1,19 @@
+package com.example.fitnessfirst.home.ui.bmi;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class NotificationsViewModel extends ViewModel {
+
+    private final MutableLiveData<String> mText;
+
+    public NotificationsViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is notifications fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
