+++ app/src/main/java/com/example/fitnessfirst/home/ui/exercise/DashboardViewModel.java
@@ -1,0 +1,19 @@
+package com.example.fitnessfirst.home.ui.exercise;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class DashboardViewModel extends ViewModel {
+
+    private final MutableLiveData<String> mText;
+
+    public DashboardViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is dashboard fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
