+++ app/src/main/java/com/example/fitnessfirst/home/ui/exercise/activities/BackExerciseActivity.java
@@ -1,0 +1,16 @@
+package com.example.fitnessfirst.home.ui.exercise.activities;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.fitnessfirst.R;
+
+public class BackExerciseActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_back_exercise);
+    }
+}
