+++ app/src/main/java/com/example/fitnessfirst/model/Todo.java
@@ -1,0 +1,10 @@
+package com.example.fitnessfirst.model;
+
+
+public class Todo {
+
+    private String name;
+
+
+
+}
