+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,24 +7,24 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.ali.fitnessfirst.home.ui.home.HomeFragment"
+        android:name="com.example.fitnessfirst.home.ui.home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.ali.fitnessfirst.home.ui.exercise.ExerciseFragment"
+        android:name="com.example.fitnessfirst.home.ui.exercise.ExerciseFragment"
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.ali.fitnessfirst.home.ui.bmi.BmiFragment"
+        android:name="com.example.fitnessfirst.home.ui.bmi.BmiFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
     <fragment
         android:id="@+id/navigation_todo"
-        android:name="com.ali.fitnessfirst.home.ui.todo.TodoFragment"
+        android:name="com.example.fitnessfirst.home.ui.todo.TodoFragment"
         android:label="@string/title_todo"
         tools:layout="@layout/fragment_todo" />
 </navigation>
